Fixes dom-fiber tests
