+++ packages/@react-facet/dom-fiber-testing-library/src/index.ts
@@ -1,4 +1,1 @@
-import { setupAct } from './setupAct'
 export * from './renderer'
-
-setupAct()
