+++ packages/@react-facet/dom-fiber/src/setupHostConfig.spec.tsx
@@ -1,29 +1,9 @@
 import { createFacet, Facet } from '@react-facet/core'
-import React, { ReactElement, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { Fiber } from 'react-reconciler'
-import { createFiberRoot } from './createFiberRoot'
-import { createReconciler } from './createReconciler'
 import { setupHostConfig } from './setupHostConfig'
+import { act, render, root } from './testSetup'
 import { InputType, ElementContainer, ElementProps, Props } from './types'
-
-document.body.innerHTML = `<div id="root"></div>`
-
-const reconcilerInstance = createReconciler()
-const root = document.getElementById('root') as HTMLElement
-const fiberRoot = createFiberRoot(reconcilerInstance)(root)
-
-/**
- * Render function local to testing that shared the same instance of the reconciler.
- *
- * This is needed otherwise React complains that we are sharing a context across different renderers.
- */
-const render = function render(ui: ReactElement) {
-  reconcilerInstance.updateContainer(ui, fiberRoot, null, () => {})
-}
-
-afterEach(() => {
-  reconcilerInstance.updateContainer(null, fiberRoot, null, () => {})
-})
 
 jest.useFakeTimers()
 
@@ -860,7 +840,9 @@ describe('update', () => {
     render(<TestComponent />)
     expect(root).toContainHTML('<div id="root"><div>Goodbye World</div></div>')
 
-    jest.runAllTimers()
+    act(() => {
+      jest.runAllTimers()
+    })
     expect(root).toContainHTML('<div id="root"><div>Hello World</div></div>')
   })
 
@@ -880,7 +862,9 @@ describe('update', () => {
     render(<TestComponent />)
     expect(root).toContainHTML('<div id="root"><div class="goodbye">Hello World</div></div>')
 
-    jest.runAllTimers()
+    act(() => {
+      jest.runAllTimers()
+    })
     expect(root).toContainHTML('<div id="root"><div class="hello">Hello World</div></div>')
   })
 
@@ -900,7 +884,9 @@ describe('update', () => {
     render(<TestComponent />)
     expect(root).toContainHTML('<div id="root"><div id="goodbye">Hello World</div></div>')
 
-    jest.runAllTimers()
+    act(() => {
+      jest.runAllTimers()
+    })
     expect(root).toContainHTML('<div id="root"><div id="hello">Hello World</div></div>')
   })
 
@@ -920,7 +906,9 @@ describe('update', () => {
     render(<TestComponent />)
     expect(root).toContainHTML('<div id="root"><div style="background: yellow;">Hello World</div></div>')
 
-    jest.runAllTimers()
+    act(() => {
+      jest.runAllTimers()
+    })
     expect(root).toContainHTML('<div id="root"><div style="background: red;">Hello World</div></div>')
   })
 
@@ -942,7 +930,9 @@ describe('update', () => {
     // jsdom is adding the http://localhost
     expect(img && img.src).toEqual('http://localhost/frita.png')
 
-    jest.runAllTimers()
+    act(() => {
+      jest.runAllTimers()
+    })
     img = document.getElementsByClassName('image')[0] as HTMLImageElement | undefined
     // jsdom is adding the http://localhost
     expect(img && img.src).toEqual('http://localhost/banana.png')
@@ -964,7 +954,9 @@ describe('update', () => {
     render(<TestComponent />)
     expect(root?.innerHTML ?? '').toBe('<a href="old"></a>')
 
-    jest.runAllTimers()
+    act(() => {
+      jest.runAllTimers()
+    })
     expect(root?.innerHTML ?? '').toBe('<a href="new"></a>')
   })
 
@@ -984,7 +976,9 @@ describe('update', () => {
     render(<TestComponent />)
     expect(root?.innerHTML ?? '').toBe('<a target="old"></a>')
 
-    jest.runAllTimers()
+    act(() => {
+      jest.runAllTimers()
+    })
     expect(root?.innerHTML ?? '').toBe('<a target="new"></a>')
   })
 
@@ -1000,9 +994,13 @@ describe('update', () => {
 
     render(<MockComponent />)
     expect(root?.innerHTML ?? '').toBe('<input value="foo">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input value="bar">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input>')
   })
 
@@ -1025,9 +1023,13 @@ describe('update', () => {
     }
     render(<MockComponent />)
     expect(root?.innerHTML ?? '').toBe('<input type="text"><input type="password"><input type="button">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input type="button"><input type="text"><input type="password">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input><input><input>')
   })
 
@@ -1044,11 +1046,17 @@ describe('update', () => {
 
     render(<MockComponent />)
     expect(root?.innerHTML ?? '').toBe('<input disabled="">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input>')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input disabled="">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input>')
   })
 
@@ -1065,11 +1073,17 @@ describe('update', () => {
 
     render(<MockComponent />)
     expect(root?.innerHTML ?? '').toBe('<input maxlength="1">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input>')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input maxlength="2">')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<input>')
   })
 
@@ -1086,11 +1100,17 @@ describe('update', () => {
 
     render(<MockComponent />)
     expect(root?.innerHTML ?? '').toBe('<div data-droppable=""></div>')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<div></div>')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<div data-droppable=""></div>')
-    jest.advanceTimersByTime(1)
+    act(() => {
+      jest.advanceTimersByTime(1)
+    })
     expect(root?.innerHTML ?? '').toBe('<div></div>')
   })
 
@@ -1108,11 +1128,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<path d="M0,0 L0,10 Z"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path d="M0,10 L0,10 Z"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path d="M0,0 L0,10 Z"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path></path>')
     })
 
@@ -1129,11 +1155,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<path fill="#ff0000"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path fill="#00ff00"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path fill="#ff0000"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path></path>')
     })
 
@@ -1150,11 +1182,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<path stroke="#ff0000"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path stroke="#00ff00"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path stroke="#ff0000"></path>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<path></path>')
     })
 
@@ -1171,11 +1209,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<svg width="500"></svg>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<svg width="600"></svg>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<svg width="500"></svg>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<svg></svg>')
     })
 
@@ -1192,11 +1236,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<svg height="500"></svg>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<svg height="600"></svg>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<svg height="500"></svg>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<svg></svg>')
     })
 
@@ -1213,11 +1263,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<rect x="500"></rect>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<rect x="600"></rect>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<rect x="500"></rect>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<rect></rect>')
     })
 
@@ -1234,11 +1290,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<rect y="500"></rect>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<rect y="600"></rect>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<rect y="500"></rect>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<rect></rect>')
     })
 
@@ -1255,11 +1317,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<circle cx="500"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle cx="600"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle cx="500"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle></circle>')
     })
 
@@ -1276,11 +1344,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<circle cy="500"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle cy="600"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle cy="500"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle></circle>')
     })
 
@@ -1297,11 +1371,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<circle r="500"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle r="600"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle r="500"></circle>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<circle></circle>')
     })
 
@@ -1318,11 +1398,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<ellipse rx="500"></ellipse>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<ellipse rx="600"></ellipse>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<ellipse rx="500"></ellipse>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<ellipse></ellipse>')
     })
 
@@ -1339,11 +1425,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<ellipse ry="500"></ellipse>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<ellipse ry="600"></ellipse>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<ellipse ry="500"></ellipse>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<ellipse></ellipse>')
     })
 
@@ -1360,11 +1452,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<line x1="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line x1="600"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line x1="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line></line>')
     })
 
@@ -1381,11 +1479,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<line x2="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line x2="600"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line x2="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line></line>')
     })
 
@@ -1402,11 +1506,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<line y1="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line y1="600"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line y1="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line></line>')
     })
 
@@ -1423,11 +1533,17 @@ describe('update', () => {
 
       render(<MockComponent />)
       expect(root?.innerHTML ?? '').toBe('<line y2="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line y2="600"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line y2="500"></line>')
-      jest.advanceTimersByTime(1)
+      act(() => {
+        jest.advanceTimersByTime(1)
+      })
       expect(root?.innerHTML ?? '').toBe('<line></line>')
     })
   })
@@ -1536,7 +1652,9 @@ describe('update', () => {
 
       firstOnClick.mockClear()
       secondOnClick.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('click'))
       expect(firstOnClick).not.toHaveBeenCalled()
@@ -1550,7 +1668,9 @@ describe('update', () => {
 
       firstOnFocus.mockClear()
       secondOnFocus.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('focus'))
       expect(firstOnFocus).not.toHaveBeenCalled()
@@ -1564,7 +1684,9 @@ describe('update', () => {
 
       firstOnBlur.mockClear()
       secondOnBlur.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('blur'))
       expect(firstOnBlur).not.toHaveBeenCalled()
@@ -1578,7 +1700,9 @@ describe('update', () => {
 
       firstOnMouseDown.mockClear()
       secondOnMouseDown.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('mousedown'))
       expect(firstOnMouseDown).not.toHaveBeenCalled()
@@ -1592,7 +1716,9 @@ describe('update', () => {
 
       firstOnMouseUp.mockClear()
       secondOnMouseUp.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('mouseup'))
       expect(firstOnMouseUp).not.toHaveBeenCalled()
@@ -1606,7 +1732,9 @@ describe('update', () => {
 
       firstOnTouchStart.mockClear()
       secondOnTouchStart.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('touchstart'))
       expect(firstOnTouchStart).not.toHaveBeenCalled()
@@ -1620,7 +1748,9 @@ describe('update', () => {
 
       firstOnTouchMove.mockClear()
       secondOnTouchMove.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('touchmove'))
       expect(firstOnTouchMove).not.toHaveBeenCalled()
@@ -1634,7 +1764,9 @@ describe('update', () => {
 
       firstOnTouchEnd.mockClear()
       secondOnTouchEnd.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('touchend'))
       expect(firstOnTouchEnd).not.toHaveBeenCalled()
@@ -1648,7 +1780,9 @@ describe('update', () => {
 
       firstOnMouseEnter.mockClear()
       secondOnMouseEnter.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('mouseenter'))
       expect(firstOnMouseEnter).not.toHaveBeenCalled()
@@ -1662,7 +1796,9 @@ describe('update', () => {
 
       firstOnMouseLeave.mockClear()
       secondOnMouseLeave.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('mouseleave'))
       expect(firstOnMouseLeave).not.toHaveBeenCalled()
@@ -1676,7 +1812,9 @@ describe('update', () => {
 
       firstOnKeyPress.mockClear()
       secondOnKeyPress.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('keypress'))
       expect(firstOnKeyPress).not.toHaveBeenCalled()
@@ -1690,7 +1828,9 @@ describe('update', () => {
 
       firstOnKeyDown.mockClear()
       secondOnKeyDown.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('keydown'))
       expect(firstOnKeyDown).not.toHaveBeenCalled()
@@ -1704,7 +1844,9 @@ describe('update', () => {
 
       firstOnKeyUp.mockClear()
       secondOnKeyUp.mockClear()
-      jest.runAllTimers()
+      act(() => {
+        jest.runAllTimers()
+      })
 
       div.dispatchEvent(new Event('keyup'))
       expect(firstOnKeyUp).not.toHaveBeenCalled()
